+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 [Full changelog](https://github.com/mozilla-ion/ion-core-addon/compare/v0.6.1...main)
 
 * #237: Remove reliance on the Firefox Pioner ID pref.
+* #237: Allow using the Rally.js support library without a bundler.
 
 # v0.6.1 (2020-12-09)